SED-1306-Tests and Group classes in folders
