+++ step-framework-parent/step-framework-timeseries/src/main/java/step/core/timeseries/TimeSeriesIngestionPipeline.java
@@ -3,6 +3,9 @@ package step.core.timeseries;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import step.core.collections.Collection;
+import step.core.timeseries.bucket.Bucket;
+import step.core.timeseries.bucket.BucketAttributes;
+import step.core.timeseries.bucket.BucketBuilder;
 
 import java.io.Closeable;
 import java.util.Map;
