+++ step-framework-parent/step-framework-timeseries/src/main/java/step/core/timeseries/aggregation/TimeSeriesAggregationQueryBuilder.java
@@ -1,4 +1,4 @@
-package step.core.timeseries;
+package step.core.timeseries.aggregation;
 
 import step.core.collections.Filter;
 import step.core.collections.Filters;
