+++ step-framework-parent/step-framework-timeseries/src/main/java/step/core/timeseries/query/TimeSeriesQuery.java
@@ -17,7 +17,7 @@
  * along with STEP.  If not, see <http://www.gnu.org/licenses/>.
  ******************************************************************************/
 
-package step.core.timeseries;
+package step.core.timeseries.query;
 
 import step.core.collections.Filter;
