+++ step-framework-parent/step-framework-timeseries/src/main/java/step/core/timeseries/query/TimeSeriesQueryBuilder.java
@@ -1,4 +1,4 @@
-package step.core.timeseries;
+package step.core.timeseries.query;
 
 import step.core.collections.Filter;
 import step.core.collections.Filters;
