+++ step-framework-parent/step-framework-timeseries/src/test/java/step/core/timeseries/TimeSeriesAggergationQueryTest.java
@@ -4,6 +4,8 @@ import org.junit.Assert;
 import org.junit.Test;
 import step.core.collections.inmemory.InMemoryCollection;
 import step.core.ql.OQLFilterBuilder;
+import step.core.timeseries.aggregation.TimeSeriesAggregationPipeline;
+import step.core.timeseries.bucket.Bucket;
 
 import java.util.Map;
 import java.util.Set;
